backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "psycopg[binary]>=3.2.6",
]
